Move elasticsearch querying logic to separate file
