+++ kghub_downloader/download_utils.py
@@ -1,18 +1,14 @@
-import json
 import logging
-import os
 import pathlib
 from typing import List, Optional
 from urllib.parse import urlparse
 
-import compress_json  # type: ignore
-import elasticsearch
-import elasticsearch.helpers
 import yaml
 from tqdm.auto import tqdm  # type: ignore
 
-from kghub_downloader.model import DownloadableResource
 from kghub_downloader import upload, schemas
+from kghub_downloader.elasticsearch import download_from_elastic_search
+from kghub_downloader.model import DownloadableResource
 
 # from compress_json import compress_json
 
@@ -79,11 +75,13 @@ def download_from_yaml(
 
         # Download file
         if item.api is not None:
-            download_from_api(item, outfile_path.name)
+            if item.api == "elasticsearch":
+                download_from_elastic_search(item, outfile_path.name)
+            else:
+                raise RuntimeError(f"API {item.api} not supported")
             continue
 
         schema = urlparse(item.expanded_url)
-
         download_fn = schemas.available_schemas.get(schema.scheme, None)
 
         if download_fn is None:
@@ -96,65 +94,3 @@ def download_from_yaml(
             upload.mirror_to_bucket(outfile_path, mirror, item.path)
 
     return None
-
-
-def download_from_api(yaml_item, outfile) -> None:
-    """
-
-    Args:
-        yaml_item: item to be download, parsed from yaml
-        outfile: where to write out file
-
-    Returns:
-
-    """
-    if yaml_item.api == "elasticsearch":
-        es_conn = elasticsearch.Elasticsearch(hosts=[yaml_item.url])
-        # FIXME: Validate query file and index parameters exist
-        query_data = compress_json.local_load(
-            os.path.join(os.getcwd(), yaml_item.query_file)
-        )
-        records = elastic_search_query(
-            es_conn, index=yaml_item.index, query=query_data
-        )
-        with open(outfile, "w") as output:
-            json.dump(records, output)
-        return None
-    else:
-        raise RuntimeError(f"API {yaml_item.api} not supported")
-
-
-def elastic_search_query(
-    es_connection,
-    index,
-    query,
-    scroll: str = "1m",
-    request_timeout: int = 60,
-    preserve_order: bool = True,
-):
-    """Fetch records from the given URL and query parameters.
-
-    Args:
-        es_connection: elastic search connection
-        index: the elastic search index for query
-        query: query
-        scroll: scroll parameter passed to elastic search
-        request_timeout: timeout parameter passed to elastic search
-        preserve_order: preserve order param passed to elastic search
-    Returns:
-        All records for query
-    """
-    records = []
-    results = elasticsearch.helpers.scan(
-        client=es_connection,
-        index=index,
-        scroll=scroll,
-        request_timeout=request_timeout,
-        preserve_order=preserve_order,
-        query=query,
-    )
-
-    for item in tqdm(results, desc="querying for index: " + index):
-        records.append(item)
-
-    return records
